security: hardening pass #5 (53 fixes) + static audit follow-up (11 fixes)
  Full-scale security audit addressing 64 total vulnerabilities across crypto,
  encoding, I/O, AI tier, and compliance subsystems.

  Pass #5 (53 fixes):
  - 8 CRITICAL: constant-time GHASH, GCM counter overflow, RLE resize,
    dictionary error reporting, BYTE_ARRAY bounds, INT4 sign extension,
    verify_chain early return
  - 18 HIGH: secure key zeroing, move-only ciphers, CSPRNG hardening,
    overflow guards, typed statistics, compliance configurability
  - 18 MEDIUM: X25519 constant-time, TLV caps, Thrift nesting, Arrow caps,
    bloom filter seed, writer validation, SHA-256 FIPS vector
  - 9 LOW: Snappy 64-bit, Thrift errors, GCM IV, mmap pre-validation

  Static audit follow-up (11 fixes):
  - Page CRC-32 (Parquet spec), mmap parity gaps, reader row_group OOB,
    Z-Order column count validation, Float16 shift UB + unaligned casts,
    feature flush data loss, getrandom EINTR, delta zigzag unsigned shift,
    statistics typed merge, compliance silent skip errors, WAL fsync checks

  29 new tests (394 -> 423). Standards: NIST SP 800-38D, FIPS 180-4,
  RFC 7748, MiFID II RTS 24, EU AI Act Art.12/13/19.

+++ include/signet/encoding/delta.hpp
@@ -77,7 +77,8 @@ inline constexpr size_t VALUES_PER_MINIBLOCK    = DEFAULT_BLOCK_SIZE / DEFAULT_M
 /// @return   The zigzag-encoded unsigned 64-bit value.
 /// @see zigzag_decode
 [[nodiscard]] inline uint64_t zigzag_encode(int64_t n) {
-    return static_cast<uint64_t>((n << 1) ^ (n >> 63));
+    // Cast to unsigned before left shift to avoid signed overflow UB (CWE-190)
+    return (static_cast<uint64_t>(n) << 1) ^ static_cast<uint64_t>(n >> 63);
 }
 
 /// Zigzag-encode a signed 32-bit integer to an unsigned representation.
@@ -353,7 +354,8 @@ inline void bit_unpack_values(const uint8_t* src,
         // if min_delta <= 0 (non-negative result), else just 0.
         std::vector<uint64_t> adjusted(DEFAULT_BLOCK_SIZE);
         for (size_t i = 0; i < block_remaining; ++i) {
-            adjusted[i] = static_cast<uint64_t>(block_deltas[i] - min_delta);
+            // Use unsigned arithmetic to avoid signed overflow UB
+            adjusted[i] = static_cast<uint64_t>(block_deltas[i]) - static_cast<uint64_t>(min_delta);
         }
         // Pad positions: store 0 (the padding deltas are 0, so adjusted = 0 - min_delta)
         // Cast to unsigned before negation to avoid signed overflow UB when min_delta == INT64_MIN
